Cleaned up unused code
