+++ app/src/main/java/com/ferhatozcelik/jetpackcomposetemplate/ui/home/MainScreen.kt
@@ -10,7 +10,6 @@ import androidx.compose.material3.Button
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
@@ -18,7 +17,6 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavHostController
@@ -30,22 +28,18 @@ import com.ferhatozcelik.jetpackcomposetemplate.data.repository.CurrencyConversi
 import com.ferhatozcelik.jetpackcomposetemplate.navigation.Screen
 import com.ferhatozcelik.jetpackcomposetemplate.ui.home.components.CatSwitch
 import com.ferhatozcelik.jetpackcomposetemplate.ui.home.components.LoadingScreen
+import com.ferhatozcelik.jetpackcomposetemplate.ui.theme.LightDarkModePreview
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import javax.inject.Inject
-import kotlin.random.Random
 
-
-// TODO: add option for big and small views
-// TODO: add change currency bar option
 @SuppressLint("UnusedMaterialScaffoldPaddingParameter")
 @Composable
 fun MainScreen(
     viewModel: HomeViewModel = hiltViewModel(),
     navController: NavHostController
 ) {
-
     Column(
         modifier = Modifier
             .fillMaxSize()
@@ -55,7 +49,6 @@ fun MainScreen(
     ) {
 
         val isUsdChecked = viewModel.isUsdSelected.collectAsState()
-
 
         when (val uiState = viewModel.uiState.collectAsState().value) {
             is UiState.Loading -> {
@@ -67,39 +60,19 @@ fun MainScreen(
             }
 
             is UiState.Success -> {
-                Row {
-                    CurrencySwitch(
-                        isUsdSelected = isUsdChecked.value,
-                        onCheckedChange = viewModel::changeCurrency
-                    )
-                }
-                CryptoListScreen(cryptoList = uiState.cryptoDataList)
+                CatSwitch(
+                    checked = isUsdChecked.value,
+                    onCheckedChange = viewModel::changeCurrency
+                )
+                CryptoList(cryptoList = uiState.cryptoDataList)
             }
         }
-
 
         Button(onClick = {
             navController.navigate(Screen.Detail.route + "/123")
         }) {
             Text(text = "Go to Detail")
         }
-    }
-}
-
-@Composable
-fun CurrencySwitch(isUsdSelected: Boolean, onCheckedChange: (Boolean) -> Unit) {
-    Row(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp),
-        horizontalArrangement = Arrangement.Center,
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        CatSwitch(
-            checked = isUsdSelected,
-            onCheckedChange = onCheckedChange,
-        )
-
     }
 }
 
@@ -206,9 +179,7 @@ fun CryptoItem(crypto: CryptoData) {
 }
 
 @Composable
-fun CryptoListScreen(cryptoList: List<CryptoData>) {
-    //val cryptoList = generateMockCryptoDataList(20) // Generate 20 mock items
-
+fun CryptoList(cryptoList: List<CryptoData>) {
     LazyColumn(modifier = Modifier.padding(16.dp)) {
         items(cryptoList) { crypto ->
             CryptoItem(crypto)
@@ -216,62 +187,6 @@ fun CryptoListScreen(cryptoList: List<CryptoData>) {
     }
 }
 
-
-fun generateMockCryptoDataList(count: Int = 20, isUsd: Boolean = true): List<CryptoData> {
-    val symbols = listOf("BTCINR", "ETHINR", "XRPINR", "LTCINR", "DOGEINR")
-    val baseAssets = listOf("BTC", "ETH", "XRP", "LTC", "DOGE")
-    val quoteAssets = listOf("INR")
-
-    var currencyMultiplier = 10.0
-
-    if (isUsd) {
-        currencyMultiplier = 100.0
-    }
-
-    return List(count) {
-        CryptoData(
-            symbol = symbols.random(),
-            baseAsset = baseAssets.random(),
-            quoteAsset = quoteAssets.random(),
-            openPrice = String.format(
-                "%.2f",
-                Random.nextDouble(10.0 * currencyMultiplier, 100.0 * currencyMultiplier)
-            ),
-            lowPrice = String.format(
-                "%.2f",
-                Random.nextDouble(50.0 * currencyMultiplier, 900.0 * currencyMultiplier)
-            ),
-            highPrice = String.format(
-                "%.2f",
-                Random.nextDouble(110.0 * currencyMultiplier, 1100.0 * currencyMultiplier)
-            ),
-            lastPrice = String.format(
-                "%.2f",
-                Random.nextDouble(80.0 * currencyMultiplier, 1000.0 * currencyMultiplier)
-            ),
-            volume = String.format(
-                "%.2f",
-                Random.nextDouble(10.0 * currencyMultiplier, 100.0 * currencyMultiplier)
-            ),
-            bidPrice = String.format(
-                "%.2f",
-                Random.nextDouble(70.0 * currencyMultiplier, 950.0 * currencyMultiplier)
-            ),
-            askPrice = String.format(
-                "%.2f",
-                Random.nextDouble(90.0 * currencyMultiplier, 1050.0 * currencyMultiplier)
-            ),
-            at = System.currentTimeMillis() - Random.nextLong(
-                0,
-                86400000
-            ) // Timestamp within the last 24 hours
-        )
-    }
-
-
-}
-
-// Mock ViewModel for Preview
 class PreviewHomeViewModel @Inject constructor(
     cryptoCoinValueRepository: CryptoCoinValueRepository,
     conversionRepository: CurrencyConversionRepository,
@@ -287,21 +202,19 @@ class PreviewHomeViewModel @Inject constructor(
 
 }
 
-@Preview(showBackground = true)
+@LightDarkModePreview
 @Composable
-fun MainScreenPreview() {
+fun MainScreenPreviewLoading() {
     MaterialTheme {
-        Surface {
-            MainScreen(
-                viewModel = PreviewHomeViewModel(
-                    cryptoCoinValueRepository = PreviewCryptoCoinValueRepository(cryptoCoinApi = PreviewCryptoCoinValueApi()),
-                    conversionRepository = PreviewCurrencyConversionRepository(
-                        previewCurrencyConversionApi = PreviewCurrencyConversionApi()
-                    ),
-                    userPreferencesDao = PreviewUserPreferencesDao()
+        MainScreen(
+            viewModel = PreviewHomeViewModel(
+                cryptoCoinValueRepository = PreviewCryptoCoinValueRepository(cryptoCoinApi = PreviewCryptoCoinValueApi()),
+                conversionRepository = PreviewCurrencyConversionRepository(
+                    previewCurrencyConversionApi = PreviewCurrencyConversionApi()
                 ),
-                navController = rememberNavController()
-            )
-        }
+                userPreferencesDao = PreviewUserPreferencesDao()
+            ),
+            navController = rememberNavController()
+        )
     }
 }
